fix(cli): keep the sync end-to-end tests off the Windows type-check
CI type-checks the commands package for Windows (go test -c, GOOS=windows).
The new end-to-end tests import the agent service so the real client and
the real server are exercised together, and that import reaches Linux-only
agent code (audio/pipewire uses syscall.Stat_t), which does not
cross-compile.

Split them into their own file behind //go:build !windows and leave the
platform-independent spec parsing where it was. A Windows CLI cannot host
an agent, so nothing is lost by not type-checking these there.
